requirements.txt
Flask==2.0.1
line_bot_sdk==2.3.0
python-dotenv==0.21.0
requests==2.25.1
transitions==0.9.0
graphviz==0.20.1
pygraphviz==1.7